fix clicks on system messages
